Extract incomplete-run resume machinery to logscan_incomplete_resume
Sprint 4, PR #2.  Splits modules/logscan_progress.py into two
cleanly-scoped modules along a natural thematic seam.

BEFORE: logscan_progress.py mixed together:
  - progress-config helpers (load, run_order, library_list)
  - snapshot builders (incomplete_progress, completed_progress)
  - a big pile of 'was this run cut short, can we resume it?'
    logic that was 65% of the file's line count

AFTER: The 8 incomplete-run/resume functions move to a new
module, logscan_progress keeps the smaller cluster.

Public API in modules/logscan_incomplete_resume:
- analyze_incomplete_log_for_resume(log_path, ...)
    THE BIG ONE (194 lines).  Parses a partial log, extracts
    run-command + processed libraries + progress snapshot,
    returns a rich resume-suggestion dict.
- build_incomplete_run_from_cache_entry(log_path, ...)
    Fast path from a logscan cache entry.
- build_incomplete_resume_cache_fields(log_path, ...)
    Extracts JUST the fields to persist back into the cache.
- build_incomplete_log_fallback(log_path, ...)
    Minimal placeholder from log-mtime when analysis fails.
- get_logscan_incomplete_runs(limit=100, ...)
- get_logscan_incomplete_run(run_key, ...)
- get_incomplete_resume_runs(limit=25, ...)
- build_latest_incomplete_resume_hint()

This is a PURE MOVE.  No parser logic changed, no branches
reordered.  Every payload field is byte-identical.

Dependency direction: logscan_incomplete_resume -> logscan_progress
(one way).  The new module imports build_incomplete_progress_snapshot
and get_progress_library_list back from logscan_progress, which
stays where it is.  No circular imports possible.

quickstart.py's 13-function import block splits along the same
seam: 5 imports stay pointed at logscan_progress, 8 move to
logscan_incomplete_resume.  The _underscore re-export names are
preserved exactly so all test call sites (qs_module._analyze_...)
keep working.

Ruff auto-cleaned the now-unused imports in logscan_progress.py
(most of the logscan_resume formatters, hashlib, etc).

Impact:
- modules/logscan_progress.py: 872 -> 310 (-562 / -64.4%)
- modules/logscan_incomplete_resume.py: NEW 618 lines
- quickstart.py: +2 lines (split one import into two)

All 1284 tests pass.  Ruff + black clean.
